refactor(client): 抽取 pinned 虚拟 tab 注入与动作包装到 sidebar/use-pinned-tabs.ts
Sidebar.tsx 拆分第四块：activePinnedTabId/pinnedRevision 状态、
pinnedEntries/pinnedVirtualTabs/augmentedTree 三个 memo（含
exhaustive-deps 豁免注释随迁）与 wrappedActions（虚拟 id 路由回
HOME 会话的 reduceFor + ptyClose 语义）原样迁出，注释逐字保留。
调用点移到 actions memo 之后（hook 无 effect，仅 useState/useMemo，
位置调整不改变行为）。

行为不变：memo 依赖数组逐字相同；wrappedActions 的 actions 入参即
原先闭包捕获的同一个 memo 产物。
